% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getAffectedRCW.R
\name{getAffectedRCW}
\alias{getAffectedRCW}
\title{Link bills to Revised Code of Washington (RCW)}
\usage{
getAffectedRCW(biennium, billId, paired = TRUE, as.xml = FALSE)
}
\arguments{
\item{biennium}{either a string or character vector of the form "XXXX-YY"}

\item{billId}{a string or vector of character strings of the format
"XX YYYY", where XX is the prefix (HB, SB, etc.) and YYYY is the
bill number}

\item{paired}{when TRUE, assumes that equal length vectors represent paired data and
will be treated as such. Set to FALSE to generate an NxN grid of input arguments}

\item{as.xml}{set to TRUE for raw XML (defaults to FALSE)}
}
\value{
By default, returns a dataframe. If \code{as.xml = TRUE}, then
returns the raw XML
}
\description{
Get a listing of all RCW citations affected by a given bill
}
\examples{
## usage for a single bill case, XML form
getAffectedRCW("2005-06", "HB 1427", as.xml = TRUE)

## generates a dataframe of affected codes from all bills in 2007-08
bills <- getLegislationByYear(c("2007","2008"))

\dontrun{codesAffected <- getAffectedRCW("2007-08", bills$BillId)}
}
