% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getAmendments.R
\name{getAmendments}
\alias{getAmendments}
\title{Get amendments to a bill}
\usage{
getAmendments(biennium, billNumber, paired = TRUE, as.xml = FALSE)
}
\arguments{
\item{biennium}{either a string or character vector of the form "XXXX-YY"}

\item{billNumber}{an integer or character string or vector containing
the bill numbers to retrieve}

\item{paired}{when TRUE, assumes that equal length vectors represent paired data and
will be treated as such. Set to FALSE to generate an NxN grid of input arguments}

\item{as.xml}{set to TRUE for raw XML (defaults to FALSE)}
}
\value{
By default, returns a dataframe. If \code{as.xml = TRUE}, then
returns the raw XML
}
\description{
Get a list of all proposed amendments (accepted and rejected) on the bill,
including the URL to the amendment text
}
\examples{
getAmendments("2007-08", "1001", as.xml = FALSE)

## get amendments for a specific set of bills
years <- c("2005-06","2007-08","2007-08","2009-10")
bills <- c(1447,1219,1001,2680)

getAmendments(years, bills, as.xml = FALSE)
}
