% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getBillSponsors.R
\name{getBillSponsors}
\alias{getBillSponsors}
\title{Get sponsor information for a bill}
\usage{
getBillSponsors(biennium, billId, paired = TRUE, as.xml = FALSE)
}
\arguments{
\item{biennium}{either a string or character vector of the form "XXXX-YY"}

\item{billId}{a string or vector of character strings of the format
"XX YYYY", where XX is the prefix (HB, SB, etc.) and YYYY is the
bill number}

\item{paired}{when TRUE, assumes that equal length vectors represent paired data and
will be treated as such. Set to FALSE to generate an NxN grid of input arguments}

\item{as.xml}{set to TRUE for raw XML (defaults to FALSE)}
}
\value{
By default, returns a dataframe. If \code{as.xml = TRUE}, then
returns the raw XML
}
\description{
Get sponsor information for a bill
}
\examples{
## get the list of all sponsors on a set of bills, filtered for primary sponsorship

spons <- getBillSponsors("2007-08", c("HB 1001", "HB 1002", "HB 1003"))
spons <- subset(spons, Type == "Primary")
}
