% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getCommitteeMembers.R
\name{getCommitteeMembers}
\alias{getCommitteeMembers}
\title{Get committee members}
\usage{
getCommitteeMembers(biennium, agency, name, as.xml = FALSE)
}
\arguments{
\item{biennium}{either a string or character vector of the form "XXXX-YY"}

\item{agency}{either "House" or "Senate", or a vector of agencies}

\item{name}{name of the committee or vector of committee names
(for a list of committee names, see \code{\link{getCommittees}})}

\item{as.xml}{set to TRUE for raw XML (defaults to FALSE)}
}
\value{
By default, returns a dataframe. If \code{as.xml = TRUE}, then
returns the raw XML
}
\description{
Get committee members
}
\examples{
## get all committee members for a select number of committees and years
years <- c("2011-12","2013-14","2015-16","2017-18")
comms <- c("Education","Judiciary","Rules")

getCommitteeMembers(years,agency = "House",comms)
}
