% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getHearings.R
\name{getHearings}
\alias{getHearings}
\title{Get hearings regarding a bill}
\usage{
getHearings(biennium, billNumber, paired = TRUE, as.xml = FALSE)
}
\arguments{
\item{biennium}{either a string or character vector of the form "XXXX-YY"}

\item{billNumber}{an integer or character string or vector containing
the bill numbers to retrieve}

\item{paired}{when TRUE, assumes that equal length vectors represent paired data and
will be treated as such. Set to FALSE to generate an NxN grid of input arguments}

\item{as.xml}{set to TRUE for raw XML (defaults to FALSE)}
}
\value{
By default, returns a list of hearings. If \code{as.xml = TRUE}, then
returns the raw XML
}
\description{
Get a list of dates, locations, and descriptions, of all
committee hearings on a particular bill
}
\section{Note}{
 The resulting XML documents contain multiple
nested lists. As such, it is not currently compatible with the XML
package's conversion to the dataframe format. Future revisions
may be made in order to help with the cleaning process and potentially
allow for better compatibility with dataframes.
}

\examples{
## get hearings for all senate bills in 2011-12
bills <- getLegislationByYear(c("2011", "2012"))
billsSenate <- subset(bills, OriginalAgency == "Senate")

\dontrun{getHearings(billsSenate$Biennium, billsSenate$BillNumber, paired = TRUE)}

}
