% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getLegislationByYear.R
\name{getLegislationByYear}
\alias{getLegislationByYear}
\title{Get legislation by year}
\usage{
getLegislationByYear(year, as.xml = FALSE)
}
\arguments{
\item{year}{Character or numeric vector representing the year(s) to be
searched.}

\item{as.xml}{If TRUE, returns data in XML format. Otherwise parses data
into dataframe format.}
}
\value{
By default, returns a dataframe. If \code{as.xml = TRUE}, then
returns the raw XML
}
\description{
Get a list of all bills introduced during the year
}
\examples{
getLegislationByYear("2007")
}
