% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getLegislationSigned.R
\name{getLegislationSigned}
\alias{getLegislationSigned}
\title{Get all bills signed into law}
\usage{
getLegislationSigned(biennium, agency, paired = FALSE, as.xml = FALSE)
}
\arguments{
\item{biennium}{either a string or character vector of the form "XXXX-YY"}

\item{agency}{either "House" or "Senate", or a vector of agencies}

\item{paired}{when TRUE, assumes that equal length vectors represent paired data and
will be treated as such. Set to FALSE to generate an NxN grid of input arguments}

\item{as.xml}{set to TRUE for raw XML (defaults to FALSE)}
}
\value{
By default, returns a dataframe. If \code{as.xml = TRUE}, then
returns the raw XML
}
\description{
Get a dataframe containing all of the bills that originated in a
given chamber and were eventually signed into law
}
\examples{
## get all bills signed into law between 2007-2012
bienniums <- c("2007-08", "2009-10", "2011-12")
getLegislationSigned(bienniums, c("House", "Senate"))
}
