% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getRCWBills.R
\name{getRCWBills}
\alias{getRCWBills}
\title{Search for bills based on Revised Code (RCW) citations}
\usage{
getRCWBills(biennium, rcwCite, paired = FALSE, as.xml = FALSE)
}
\arguments{
\item{biennium}{either a string or character vector of the form "XXXX-YY"}

\item{rcwCite}{character string (or vector) for the citation (optional extensions
for title, chapter, or section) in the RCW to pull legislation
from}

\item{paired}{when TRUE, assumes that equal length vectors represent paired data and
will be treated as such. Set to FALSE to generate an NxN grid of input arguments}

\item{as.xml}{set to TRUE for raw XML (defaults to FALSE)}
}
\value{
default, returns a dataframe. If \code{as.xml = TRUE}, then
returns the raw XML
}
\description{
Get a list of all bills which reference or amend the Revised Code of
Washington (RCW)
}
\examples{
getRCWBills("2007-08", "13.40.0357")
}
