% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getSponsors.R
\name{getSponsors}
\alias{getSponsors}
\title{Get legislators}
\usage{
getSponsors(biennium, as.xml = FALSE)
}
\arguments{
\item{biennium}{either a string or character vector of the form "XXXX-YY"}

\item{as.xml}{set to TRUE for raw XML (defaults to FALSE)}
}
\value{
By default, returns a dataframe. If \code{as.xml = TRUE}, then
returns the raw XML
}
\description{
Get a list of all sponsors (all congressmembers) for a given biennium
}
\examples{
getSponsors("2007-08")
}
