% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getAffectedRCW.R
\name{getAffectedRCW}
\alias{getAffectedRCW}
\title{Link bills to Revised Code of Washington (RCW)}
\usage{
getAffectedRCW(biennium, billId, paired = TRUE, as.xml = FALSE)
}
\arguments{
\item{biennium}{Character vector representing the biennium(s) to be
searched. Each argument should take the form "XXXX-YY"}

\item{billId}{Character vector containing the bill(s) to be retrieved.
Each argument should take the form "XX YYYY", where XX
is the prefix (HB, SB, etc.) and YYYY is the bill number.}

\item{paired}{If TRUE, will assume that equal length vectors represent
paired data. Set to FALSE to generate an NxN grid of input
arguments. Applies to equal length vector inputs only.}

\item{as.xml}{If TRUE, returns data in XML format. Otherwise parses data
into dataframe format.}
}
\value{
By default, returns a dataframe. If \code{as.xml = TRUE}, then
returns the raw XML
}
\description{
Get a listing of all RCW citations affected by a given bill
}
\examples{
## usage for a single bill case, XML form
getAffectedRCW("2005-06", "HB 1427", as.xml = TRUE)

## generates a dataframe of affected codes from all bills in 2007-08
bills <- getLegislationByYear(c("2007","2008"))

\dontrun{codesAffected <- getAffectedRCW("2007-08", bills$BillId)}
}
