% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getCommittees.R
\name{getCommittees}
\alias{getCommittees}
\title{Get legislative committees}
\usage{
getCommittees(biennium, as.xml = FALSE)
}
\arguments{
\item{biennium}{Character vector representing the biennium(s) to be
searched. Each argument should take the form "XXXX-YY"}

\item{as.xml}{If TRUE, returns data in XML format. Otherwise parses data
into dataframe format.}
}
\value{
By default, returns a dataframe. If \code{as.xml = TRUE}, then
returns the raw XML
}
\description{
Get a list of all committees that were active during the biennium,
along with their respective committee code
}
\examples{
getCommittees("2007-08")
}
