% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getCurrentStatus.R
\name{getCurrentStatus}
\alias{getCurrentStatus}
\title{Get bill status}
\usage{
getCurrentStatus(biennium, billNumber, paired = TRUE, as.xml = FALSE)
}
\arguments{
\item{biennium}{either a string or character vector of the form "XXXX-YY"}

\item{billNumber}{an integer or character string or vector containing
the bill numbers to retrieve}

\item{paired}{when TRUE, assumes that equal length vectors represent paired data and
will be treated as such. Set to FALSE to generate an NxN grid of input arguments}

\item{as.xml}{set to TRUE for raw XML (defaults to FALSE)}
}
\value{
By default, returns a dataframe. If \code{as.xml = TRUE}, then
returns the raw XML
}
\description{
Get the current status of a given bill
}
\section{Note}{

This function returns the bill's status as of today. If a bill
was never passed, it lists the most recent status. To
get a bill's complete history, use \code{\link{getStatusChanges}}
}

\examples{
getCurrentStatus("2007-08", "1001")

## get final status for all bills written in 2011 or 2012
billNums <- rbind(getLegislationByYear("2011"), getLegislationByYear("2012"))
\dontrun{status <- getCurrentStatus("2011-12", billNums$billNumber)}

}
