% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getLegislation.R
\name{getLegislation}
\alias{getLegislation}
\title{Get summary information on a particular bill}
\usage{
getLegislation(biennium, billNumber, paired = TRUE, as.xml = FALSE)
}
\arguments{
\item{biennium}{Character vector representing the biennium(s) to be
searched. Each argument should take the form "XXXX-YY"}

\item{billNumber}{Character or numeric vector containing the bill number(s)
to be retrieved.}

\item{paired}{If TRUE, will assume that equal length vectors represent
paired data. Set to FALSE to generate an NxN grid of input
arguments. Applies to equal length vector inputs only.}

\item{as.xml}{If TRUE, returns data in XML format. Otherwise parses data
into dataframe format.}
}
\value{
By default, returns a dataframe. If \code{as.xml = TRUE}, then
returns the raw XML
}
\description{
Get legislative summary information for a particular bill,
including bill ID, introduction date, bill title(s), and
description
}
\examples{
getLegislation("2007-08", "1001")

\dontrun{getLegislation("2007-08", 1001:1100, as.xml = TRUE)}
}
