% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getLegislation.R
\name{getLegislation}
\alias{getLegislation}
\title{Get summary information on a particular bill}
\usage{
getLegislation(biennium, billNumber, paired = TRUE, as.xml = FALSE)
}
\arguments{
\item{biennium}{either a string or character vector of the form "XXXX-YY"}

\item{billNumber}{an integer or character string or vector containing
the bill numbers to retrieve}

\item{paired}{when TRUE, assumes that equal length vectors represent paired data and
will be treated as such. Set to FALSE to generate an NxN grid of input arguments}

\item{as.xml}{set to TRUE for raw XML (defaults to FALSE)}
}
\value{
By default, returns a dataframe. If \code{as.xml = TRUE}, then
returns the raw XML
}
\description{
Get legislative summary information for a particular bill,
including bill ID, introduction date, bill title(s), and
description
}
\examples{
getLegislation("2007-08", "1001")

\dontrun{getLegislation("2007-08", 1001:1100, as.xml = TRUE)}
}
