% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getLegislationByYear.R
\name{getLegislationByYear}
\alias{getLegislationByYear}
\title{Get legislation by year}
\usage{
getLegislationByYear(year, type = c("df", "list", "xml"))
}
\arguments{
\item{year}{Character or numeric vector representing the year(s) to be
searched.}

\item{type}{One of "df", "list", or "xml". Specifies the format for
the output.}
}
\value{
By default, returns a dataframe. If \code{as.xml = TRUE}, then
returns the raw XML
}
\description{
Get a list of all bills introduced during the year
}
\examples{
getLegislationByYear("2007")
}
