% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getLegislationByYear.R
\name{getLegislationByYear}
\alias{getLegislationByYear}
\title{Get legislation by year}
\usage{
getLegislationByYear(year, as.xml = FALSE)
}
\arguments{
\item{year}{an integer or character string or vector for the given year}

\item{as.xml}{set to TRUE for raw XML (defaults to FALSE)}
}
\value{
By default, returns a dataframe. If \code{as.xml = TRUE}, then
returns the raw XML
}
\description{
Get a list of all bills introduced during the year
}
\examples{
getLegislationByYear("2007")
}
