% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getLegislationSigned.R
\name{getLegislationSigned}
\alias{getLegislationSigned}
\title{Get all bills signed into law}
\usage{
getLegislationSigned(
  biennium,
  agency = c("House", "Senate"),
  paired = FALSE,
  as.xml = FALSE
)
}
\arguments{
\item{biennium}{Character vector representing the biennium(s) to be
searched. Each argument should take the form "XXXX-YY"}

\item{agency}{One of "House" or "Senate", or a vector with these as its
elements.}

\item{paired}{If TRUE, will assume that equal length vectors represent
paired data. Set to FALSE to generate an NxN grid of input
arguments. Applies to equal length vector inputs only.}

\item{as.xml}{If TRUE, returns data in XML format. Otherwise parses data
into dataframe format.}
}
\value{
By default, returns a dataframe. If \code{as.xml = TRUE}, then
returns the raw XML
}
\description{
Get a dataframe containing all of the bills that originated in a
given chamber and were eventually signed into law
}
\examples{
## get all bills signed into law between 2007-2012
bienniums <- c("2007-08", "2009-10", "2011-12")
getLegislationSigned(bienniums, c("House", "Senate"))
}
