% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getRCWBills.R
\name{getRCWBills}
\alias{getRCWBills}
\title{Search for bills based on Revised Code (RCW) citations}
\usage{
getRCWBills(biennium, rcwCite, paired = FALSE, as.xml = FALSE)
}
\arguments{
\item{biennium}{Character vector representing the biennium(s) to be
searched. Each argument should take the form "XXXX-YY"}

\item{rcwCite}{Character vector for the citation in the RCW to pull
legislation from. Optional extensions for title, chapter, and section
are allowed. For more information, see
\url{https://apps.leg.wa.gov/rcw/}}

\item{paired}{If TRUE, will assume that equal length vectors represent
paired data. Set to FALSE to generate an NxN grid of input
arguments. Applies to equal length vector inputs only.}

\item{as.xml}{If TRUE, returns data in XML format. Otherwise parses data
into dataframe format.}
}
\value{
default, returns a dataframe. If \code{as.xml = TRUE}, then
returns the raw XML
}
\description{
Get a list of all bills which reference or amend a particular
portion of the Revised Code of Washington (RCW)
}
\examples{
getRCWBills("2007-08", "13.40.0357")
}
