% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getRollCalls.R
\name{getRollCalls}
\alias{getRollCalls}
\title{Get roll call votes}
\usage{
getRollCalls(biennium, billNumber)
}
\arguments{
\item{biennium}{either a string or character vector of the form "XXXX-YY"}

\item{billNumber}{an integer or character string or vector containing
the bill numbers to retrieve}
}
\value{
Currently, returns an XML array containing all roll call votes.
For usage, see examples below
}
\description{
Get an XML containing roll call information for all recorded votes
on a bill
}
\section{Note}{
 This function as implemented is not in its final state.
The resulting XML document(s) is heavily nested and
typically requires further cleaning.
See examples below for some potential usages. Further
revisions will be made to help facilitate the XML parsing and
potentially allow for compatibility with dataframes.
}

\examples{
votes <- getRollCalls("2007-08", "1001") # get roll call votes

length(votes) # total number of roll call votes recorded

names(votes[[1]]) # view all possible subcategories under roll call vote #1

votes[[1]][["Motion"]][[1]] # motion title on roll call vote #1

votes[[3]][["YeaVotes"]][["Count"]] # number of yea votes on roll call vote #3

}
