% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getStatusChanges.R
\name{getStatusChanges}
\alias{getStatusChanges}
\title{Track historical progess on a bill}
\usage{
getStatusChanges(biennium, billNumber, paired = TRUE, as.xml = FALSE)
}
\arguments{
\item{biennium}{either a string or character vector of the form "XXXX-YY"}

\item{billNumber}{an integer or character string or vector containing
the bill numbers to retrieve}

\item{paired}{when TRUE, assumes that equal length vectors represent paired data and
will be treated as such. Set to FALSE to generate an NxN grid of input arguments}

\item{as.xml}{set to TRUE for raw XML (defaults to FALSE)}
}
\value{
By default, returns a dataframe. If \code{as.xml = TRUE}, then
returns the raw XML
}
\description{
Get a complete history of all status changes that occurred on a particular
bill
}
\examples{
getStatusChanges("2007-08", "1001", as.xml = FALSE)
}
